reorganized factories due to extension not working when trying to use them. switched back rpc query id to be int instead of string, for single type managment purpose: theorically an rpc id can be string too, but for simplicity all ids assigned will be int
